+++ ZigBeeNet/ZDO/Command/MatchDescriptorRequest.cs
@@ -33,12 +33,12 @@ namespace ZigBeeNet.ZDO.Command
         /**
         * InClusterList command message field.
         */
-        public List<int> InClusterList { get; set; }
+        public List<ushort> InClusterList { get; set; }
 
         /**
         * OutClusterList command message field.
         */
-        public List<int> OutClusterList { get; set; }
+        public List<ushort> OutClusterList { get; set; }
 
         /**
         * Default constructor.
@@ -74,8 +74,8 @@ namespace ZigBeeNet.ZDO.Command
             base.Deserialize(deserializer);
 
             // Create lists
-            InClusterList = new List<int>();
-            OutClusterList = new List<int>();
+            InClusterList = new List<ushort>();
+            OutClusterList = new List<ushort>();
 
             NwkAddrOfInterest = (int)deserializer.Deserialize(ZclDataType.Get(DataType.NWK_ADDRESS));
             ProfileId = (int)deserializer.Deserialize(ZclDataType.Get(DataType.UNSIGNED_16_BIT_INTEGER));
@@ -85,7 +85,7 @@ namespace ZigBeeNet.ZDO.Command
             {
                 for (int cnt = 0; cnt < inClusterCount; cnt++)
                 {
-                    InClusterList.Add((int)deserializer.Deserialize(ZclDataType.Get(DataType.CLUSTERID)));
+                    InClusterList.Add((ushort)deserializer.Deserialize(ZclDataType.Get(DataType.CLUSTERID)));
                 }
             }
 
@@ -95,7 +95,7 @@ namespace ZigBeeNet.ZDO.Command
             {
                 for (int cnt = 0; cnt < outClusterCount; cnt++)
                 {
-                    OutClusterList.Add((int)deserializer.Deserialize(ZclDataType.Get(DataType.CLUSTERID)));
+                    OutClusterList.Add((ushort)deserializer.Deserialize(ZclDataType.Get(DataType.CLUSTERID)));
                 }
             }
         }
